requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-classic
langchain-core
langchain-text-splitters
python-dotenv
sentence-transformers
faiss-cpu
requests
unstructured
python-docx
pypdf
tiktoken
google-generativeai
